Moved FutureResults class and method
moved into schedulefunctions.dart for easier finding
