+++ friend_me/lib/pages/schedule.dart
@@ -4,7 +4,7 @@ import 'package:flutter/foundation.dart';
 import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:kalender/kalender.dart';
-import 'package:friend_me/widgets/schedulefunctions.dart';
+import 'package:friend_me/widgets/schedulefunctions.dart'; 
 import 'package:friend_me/widgets/event.dart';
 import 'package:intl/intl.dart';
 import 'package:http/http.dart' as http;
@@ -94,7 +94,7 @@ class _ScheduleRouteState extends State<ScheduleRoute>
           // Future builder. waits for connection and Displays calendar on succcessful return from backend, otherwise shows error screen or loading circle appropriately
           //FutureResults holds UID and initial response from http.get.  
           body: FutureBuilder<FutureResults>(
-              future: _getResults(calendar.eventsController),
+              future: getResults(calendar.eventsController),
               builder: (context, AsyncSnapshot snapshot) {
                 if (snapshot.connectionState == ConnectionState.waiting) {
                   return const Center(
@@ -432,16 +432,3 @@ String getTitle(Event? e) {
 }
 
 //holds values for future builder
-class FutureResults{
-  final http.Response Response; 
-  final String? UID; 
-  FutureResults({required this.Response, required this.UID});
-}
-
-  Future<FutureResults> _getResults(CalendarEventsController controller) async {
-    final String? _uid = await getUsername(); 
-    print("uid: $_uid");
-    final http.Response _response = await fetchEvents(controller, _uid); 
-    print("response: ${_response.statusCode}"); 
-    return FutureResults(Response: _response, UID : _uid);
-  }
